I included the analysis of the smeared data tree in both multiplicity and S0pT1. I also added a multiplicity response matrix.
pythia8_generate-tree.ipynb:
I now include the mass of the particles in the events.root.
I added a new section for simulation of detector effects using RDataFrame. I wrote simple simulations for acceptance of TPC and TOF and for smearing pT, eta, phi utilizing the PtEtaPhiMVector. The efficiency and smearing functions are written in C++ and loaded via the ROOT's interpreter.

+++ cpp/S0pT1.cpp
@@ -3,8 +3,8 @@
 #include <ROOT/RVec.hxx> // ROOT's fast array of numbers
 
 // define the function to calculate unweighted transverse spherocity for an event
-double calc_S0pT1(const ROOT::RVec<float>& uxEvent, // const is a safety lock to ensure that the function only reads data and does not change it
-                  const ROOT::RVec<float>& uyEvent, // & tells C++ to use the original data rather than making copies in RAM
+double calc_S0pT1(const ROOT::RVec<double>& uxEvent, // const is a safety lock to ensure that the function only reads data and does not change it
+                  const ROOT::RVec<double>& uyEvent, // & tells C++ to use the original data rather than making copies in RAM
                   int NchEvent) {
 
     double minVal = 999999.0; // set minimum value to a large number, so that it gets overwritten
